refactor: ranking
