+++ src/features/ranking/components/current-user-card-prefecture.tsx
@@ -1,5 +1,5 @@
-import type { UserRanking } from "@/lib/services/ranking";
-import BaseCurrentUserCard from "./base-current-user-card";
+import type { UserRanking } from "../types/ranking-types";
+import { BaseCurrentUserCard } from "./base-current-user-card";
 import { LevelBadge } from "./ranking-level-badge";
 
 interface CurrentUserCardProps {
