+++ src/features/ranking/components/index.ts
@@ -1,3 +1,3 @@
 // エクスポート用ファイル
-export { default as RankingTop } from "./ranking-top";
+export { RankingTop } from "./ranking-top";
 export { RankingItem } from "./ranking-item";
