+++ src/features/ranking/components/ranking-icon.test.tsx
