+++ src/features/ranking/components/ranking-icon.tsx
